requirements.txt
telethon
tgcrypto
swibots
aioredis
aiofiles
python-decouple